+++ src/OpenTelemetry.Extensions.AWS/AWSXRayIdGenerator.cs
@@ -14,6 +14,7 @@
 // limitations under the License.
 // </copyright>
 
+#if !NET6_0_OR_GREATER
 using System;
 using System.Diagnostics;
 using System.Globalization;
@@ -50,7 +51,6 @@ public static class AWSXRayIdGenerator
                 {
                     var awsXRayTraceId = GenerateAWSXRayCompatibleTraceId();
 
-                    // TODO: Apply API to directly modify trace id once .NET runtime publicizes it.
                     activity.SetParentId(awsXRayTraceId, default, activity.ActivityTraceFlags);
 
                     // When not using instrumented library and creating root activity using ActivitySource.StartActivity(),
@@ -197,3 +197,4 @@ public static class AWSXRayIdGenerator
         return ActivitySamplingResult.PropagationData;
     }
 }
+#endif