+++ test/OpenTelemetry.Extensions.AWS.Tests/AWSXRayIdGeneratorTests.cs
@@ -36,7 +36,12 @@ public class AWSXRayIdGeneratorTests
             activity.Start();
 
             Assert.NotEqual(originalTraceId, activity.TraceId);
+#if NET6_0_OR_GREATER
+            // the net6.0 version of AWSXRayIdGenerator uses Activity.TraceIdGenerator, which does not change the parent ID
+            Assert.Equal(originalParentSpanId, activity.ParentSpanId);
+#else
             Assert.NotEqual(originalParentSpanId, activity.ParentSpanId);
+#endif
             Assert.Equal("0000000000000000", activity.ParentSpanId.ToHexString());
             Assert.Equal(originalTraceFlag, activity.ActivityTraceFlags);
         }
@@ -81,11 +86,13 @@ public class AWSXRayIdGeneratorTests
     [Fact]
     public void TestGenerateTraceIdForRootNodeUsingActivitySourceWithTraceIdBasedSamplerOn()
     {
+#pragma warning disable CS0618 // Type or member is obsolete
         using (Sdk.CreateTracerProviderBuilder()
                    .AddXRayTraceIdWithSampler(new TraceIdRatioBasedSampler(1.0))
                    .AddSource("TestTraceIdBasedSamplerOn")
                    .SetSampler(new TraceIdRatioBasedSampler(1.0))
                    .Build())
+#pragma warning restore CS0618 // Type or member is obsolete
         {
             using (var activitySource = new ActivitySource("TestTraceIdBasedSamplerOn"))
             {
@@ -100,11 +107,13 @@ public class AWSXRayIdGeneratorTests
     [Fact]
     public void TestGenerateTraceIdForRootNodeUsingActivitySourceWithTraceIdBasedSamplerOff()
     {
+#pragma warning disable CS0618 // Type or member is obsolete
         using (Sdk.CreateTracerProviderBuilder()
                    .AddXRayTraceIdWithSampler(new TraceIdRatioBasedSampler(0.0))
                    .AddSource("TestTraceIdBasedSamplerOff")
                    .SetSampler(new TraceIdRatioBasedSampler(0.0))
                    .Build())
+#pragma warning restore CS0618 // Type or member is obsolete
         {
             using (var activitySource = new ActivitySource("TestTraceIdBasedSamplerOff"))
             {