+++ src/OpenTelemetry.Extensions.AWS/CHANGELOG.md
@@ -21,6 +21,9 @@
   supporting ResourceBuilderExtensions extension, and migrate all detectors to
   implement OpenTelemetry.Resources.IResourceDetector
   ([#875](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/875))
+* Add a `net6.0` build with optimized trace ID generation using the new
+  `Activity.TraceIdGenerator` API.
+  ([#1096](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/1096))
 * Drop support for `AWSLambdaResourceDetector`.
   AWS Lambda Resources are detected by `OpenTelemetry.Instrumentation.AWSLambda`
   package